move state of grid in GameContext
