+++ src/client/pages/game/Game.jsx
@@ -3,11 +3,12 @@ import FlexBox from "components/flexbox/FlexBox";
 import TetrisGrid from "components/tetris/Grid";
 import { useTetrisGame } from "hooks";
 import { Link } from "react-router-dom";
+import { GameContext } from "store";
 import NextPieces from "./NextPieces";
-// import { GameContext } from "store";
 
 export default function Game() {
-  const { tetrisGrid, ...tetris } = useTetrisGame(10, 20);
+  const { state } = React.useContext(GameContext);
+  const tetris = useTetrisGame(10, 20);
 
   return (
     <FlexBox
@@ -17,14 +18,14 @@ export default function Game() {
       className="justify-center items-center"
     >
       <Link to="/">Retour accueil</Link>
+      <button onClick={tetris.toggleShortcuts}>Toggle Shortcuts</button>
       <FlexBox direction="row" className="mt-4">
         <FlexBox
           direction="col"
           width={64}
           className="justify-center align-center"
         >
-          <TetrisGrid grid={tetrisGrid} />
-          <button onClick={() => tetris.movePiece("DOWN")}>go down</button>
+          <TetrisGrid grid={state.grid} />
         </FlexBox>
         <NextPieces />
       </FlexBox>
