+++ src/client/reducers/index.js
@@ -3,4 +3,4 @@ export { default as storeReducer, initialState as initialStore } from "./store";
 export {
   default as piecesReducer,
   initialState as initialPieces,
-} from "./pieces";
+} from "./game";
